requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bson==0.5.10
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
# command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dataclasses-json==0.5.9
dbus-python==1.2.18
distro==1.7.0
# distro-info===1.1build1
Django==4.1.7
django-cors-headers==3.13.0
django-extra-keywords==1.1.2
django-jsonfield==1.4.1
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
# gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
langchain==0.0.225
langchainplus-sdk==0.0.20
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-index==0.7.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netifaces==0.11.0
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.0
openai==0.27.8
openapi-codec==1.3.2
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pydantic==1.10.11
PyGObject==3.42.1
PyJWT==2.6.0
pyparsing==2.4.7
pypng==0.20220715.0
pyrsistent==0.19.3
# python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
regex==2023.6.3
requests==2.31.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
spectacular==0.4.0
SQLAlchemy==2.0.18
sqlparse==0.4.3
# systemd-python==234
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.16
wadllib==1.3.6
Werkzeug==2.3.6
yarl==1.9.2
zipp==1.0.0
